// server/enums/supportStrategyType.ts
enum SupportStrategyType {
  LITERACY_SKILLS_DEFAULT = 'LITERACY_SKILLS_DEFAULT',
  NUMERACY_SKILLS_DEFAULT = 'NUMERACY_SKILLS_DEFAULT',
  ATTENTION_ORGANISING_TIME_DEFAULT = 'ATTENTION_ORGANISING_TIME_DEFAULT',
  LANGUAGE_COMM_SKILLS_DEFAULT = 'LANGUAGE_COMM_SKILLS_DEFAULT',
  EMOTIONS_FEELINGS_DEFAULT = 'EMOTIONS_FEELINGS_DEFAULT',
  PHYSICAL_SKILLS_DEFAULT = 'PHYSICAL_SKILLS_DEFAULT',
  SENSORY = 'SENSORY',
  MEMORY = 'MEMORY',
  PROCESSING_SPEED = 'PROCESSING_SPEED',
  GENERAL = 'GENERAL',
}

export default SupportStrategyType