// server/enums/supportStrategyCategory.ts
enum SupportStrategyCategory {
  SENSORY = 'SENSORY',
  EMOTIONS_FEELINGS = 'EMOTIONS_FEELINGS',
  PHYSICAL_SKILLS = 'PHYSICAL_SKILLS',
  LITERACY_SKILLS = 'LITERACY_SKILLS',
  NUMERACY_SKILLS = 'NUMERACY_SKILLS',
  ATTENTION_ORGANISING_TIME = 'ATTENTION_ORGANISING_TIME',
  LANGUAGE_COMM_SKILLS = 'LANGUAGE_COMM_SKILLS',
  PROCESSING_SPEED = 'PROCESSING_SPEED',
  MEMORY = 'MEMORY',
}

export default SupportStrategyCategory